Add Auction creation logic (unfinished)
+ Finish page create Auction with item name, description, starting price. The starting time input is not worked yet and the default time would be now
+ Slight modification for Auction and Auction Manager. Now createAuction() only receive Item and Seller. Other information will be taken from Seller and Item

+++ src/main/java/Branch/Item.java
@@ -12,13 +12,13 @@ public class Item extends Entity {
     }
     private Status status = Status.AVAILABLE;
 
-    public Item(String name, float startingPrice, String description) {
+    public Item(String name, double startingPrice, String description) {
         super(name);
         this.startingPrice = startingPrice;
         this.description = description;
     }
 
-    public void setStartingPrice(float startPrice) {
+    public void setStartingPrice(double startPrice) {
         this.startingPrice = startPrice;
     }
